Move all editor code in the editors adapters module
