+++ gaphor/adapters/editors.py
@@ -1,16 +1,13 @@
 """
-Adapters
+Editors.
 """
 import logging
-from zope import component
-
+import abc
 from functools import singledispatch
-from zope.interface import implementer
 
 from gaphor import UML
 from gaphor.core import inject
 from gaphor.diagram import items
-from gaphor.diagram.abc import EditorBase
 from gaphor.diagram.interfaces import Editor
 from gaphor.misc.rattr import rgetattr, rsetattr
 
@@ -35,8 +32,51 @@ def editable_slot(el):
     return el.value
 
 
+class AbstractEditor(metaclass=abc.ABCMeta):
+    """
+    Provide an interface for editing text with the TextEditTool.
+    """
+
+    @abc.abstractmethod
+    def is_editable(self, x, y):
+        """
+        Is this item editable in it's current state.
+        x, y represent the cursors (x, y) position.
+        (this method should be called before get_text() is called.
+        """
+
+    @abc.abstractmethod
+    def get_text(self):
+        """
+        Get the text to be updated
+        """
+
+    @abc.abstractmethod
+    def get_bounds(self):
+        """
+        Get the bounding box of the (current) text. The edit tool is not
+        required to do anything with this information but it might help for
+        some nicer displaying of the text widget.
+
+        Returns: a gaphas.geometry.Rectangle
+        """
+
+    @abc.abstractmethod
+    def update_text(self, text):
+        """
+        Update with the new text.
+        """
+
+    @abc.abstractmethod
+    def key_pressed(self, pos, key):
+        """
+        Called every time a key is pressed. Allows for 'Enter' as escape
+        character in single line editing.
+        """
+
+
 @Editor.register(items.CommentItem)
-class CommentItemEditor(EditorBase):
+class CommentItemEditor(AbstractEditor):
     """Text edit support for Comment item."""
 
     def __init__(self, item):
@@ -59,7 +99,7 @@ class CommentItemEditor(EditorBase):
 
 
 @Editor.register(items.NamedItem)
-class NamedItemEditor(EditorBase):
+class NamedItemEditor(AbstractEditor):
     """Text edit support for Named items."""
 
     def __init__(self, item):
@@ -85,7 +125,7 @@ class NamedItemEditor(EditorBase):
 
 
 @Editor.register(items.DiagramItem)
-class DiagramItemTextEditor(EditorBase):
+class DiagramItemTextEditor(AbstractEditor):
     """Text edit support for diagram items containing text elements."""
 
     def __init__(self, item):
@@ -120,7 +160,7 @@ class DiagramItemTextEditor(EditorBase):
 
 
 @Editor.register(items.CompartmentItem)
-class CompartmentItemEditor(EditorBase):
+class CompartmentItemEditor(AbstractEditor):
     """Text editor support for compartment items."""
 
     def __init__(self, item):
@@ -149,7 +189,7 @@ class CompartmentItemEditor(EditorBase):
 
 
 @Editor.register(items.AssociationItem)
-class AssociationItemEditor(EditorBase):
+class AssociationItemEditor(AbstractEditor):
     def __init__(self, item):
         self._item = item
         self._edit = None
@@ -192,7 +232,7 @@ class AssociationItemEditor(EditorBase):
 
 
 @Editor.register(items.ForkNodeItem)
-class ForkNodeItemEditor(EditorBase):
+class ForkNodeItemEditor(AbstractEditor):
     """Text edit support for fork node join specification."""
 
     element_factory = inject("element_factory")
