+++ gaphor/diagram/interfaces.py
@@ -26,40 +26,7 @@ class IConnect(interface.Interface):
     adapter could be written.
     """
 
-    def connect(self, handle, port):
-        """
-        Connect a line's handle to element.
-
-        Note that at the moment of the connect, handle.connected_to may point
-        to some other item. The implementor should do the disconnect of
-        the other element themselves.
-        """
-
-    def disconnect(self, handle):
-        """
-        The true disconnect. Disconnect a handle.connected_to from an
-        element. This requires that the relationship is also removed at
-        model level.
-        """
-
-    def connect_constraints(self, handle):
-        """
-        Connect a handle to the element.
-        """
-
-    def disconnect_constraints(self, handle):
-        """
-        Disconnect a line's handle from an element.
-        This is called whenever a handle is dragged.
-        """
-
-    def glue(self, handle):
-        """
-        Determine if a handle can glue to a specific element.
-
-        Returns a tuple (x, y) if the line and element may connect, None
-        otherwise.
-        """
+    pass
 
 
 # TODO: I think this should have been called Namespacing or something similar,
